requirements.txt
fastapi==0.104.1
uvicorn==0.23.2
sqlalchemy==2.0.23
pydantic==2.4.2
pydantic-settings==2.0.3
python-dotenv==1.0.0
httpx==0.25.1
jinja2==3.1.2
pytest==7.4.3
alembic==1.12.1
black==23.10.1
isort==5.12.0
python-jose==3.3.0
passlib==1.7.4
python-multipart==0.0.6
bcrypt==4.0.1 